Restructing tests
