+++ test/command/migration_test.dart
@@ -142,80 +142,7 @@ void main() {
     });
   });
 
-  group("Generating migration files", () {
-    var projectSourceDirectory = getTestProjectDirectory("initial");
-    Directory projectDirectory = new Directory("test_project");
-    var migrationDirectory =
-        new Directory.fromUri(projectDirectory.uri.resolve("migrations"));
-    var addFiles = (List<String> filenames) {
-      filenames.forEach((name) {
-        new File.fromUri(migrationDirectory.uri.resolve(name))
-            .writeAsStringSync(" ");
-      });
-    };
 
-    setUp(() async {
-      createTestProject(projectSourceDirectory, projectDirectory);
-    });
-
-    tearDown(() {
-      projectDirectory.deleteSync(recursive: true);
-    });
-
-    test("Run without pub get yields error", () async {
-      var out = await runAqueductProcess(["db", "generate"], projectDirectory);
-      expect(out != 0, true);
-    });
-
-    test("Ensure migration directory will get created on generation", () async {
-      await runPubGet(projectDirectory, offline: true);
-      expect(migrationDirectory.existsSync(), false);
-      var out = await runAqueductProcess(["db", "generate"], projectDirectory);
-      expect(out, 0);
-      expect(migrationDirectory.existsSync(), true);
-    });
-
-    test(
-        "If there are no migration files, create an initial one that validates to schema",
-        () async {
-      await runPubGet(projectDirectory, offline: true);
-
-      // Putting a non-migration file in there to ensure that this doesn't prevent from being ugpraded
-      migrationDirectory.createSync();
-      addFiles(["notmigration.dart"]);
-
-      await runAqueductProcess(["db", "generate"], projectDirectory);
-      var out = await runAqueductProcess(["db", "validate"], projectDirectory);
-      expect(out, 0);
-    });
-
-    test("If there is already a migration file, create an upgrade file",
-        () async {
-      await runPubGet(projectDirectory, offline: true);
-
-      await runAqueductProcess(["db", "generate"], projectDirectory);
-      await runAqueductProcess(["db", "generate"], projectDirectory);
-
-      expect(
-          migrationDirectory
-              .listSync()
-              .where((fse) => !fse.uri.pathSegments.last.startsWith(".")),
-          hasLength(2));
-      expect(
-          new File.fromUri(migrationDirectory.uri
-                  .resolve("00000001_Initial.migration.dart"))
-              .existsSync(),
-          true);
-      expect(
-          new File.fromUri(migrationDirectory.uri
-                  .resolve("00000002_Unnamed.migration.dart"))
-              .existsSync(),
-          true);
-
-      var out = await runAqueductProcess(["db", "validate"], projectDirectory);
-      expect(out, 0);
-    });
-  });
 
   group("Validating", () {
     var projectSourceDirectory = getTestProjectDirectory("initial");
